refactor(j-s): move SignedVerdictOverviewForm to SignedVerdictOverview
