+++ apps/judicial-system/web/src/routes/Shared/SignedVerdictOverview/SignedVerdictOverview.tsx
@@ -1,4 +1,10 @@
-import React, { ReactNode, useContext, useEffect, useState } from 'react'
+import React, {
+  ReactNode,
+  useCallback,
+  useContext,
+  useEffect,
+  useState,
+} from 'react'
 import { AnimatePresence, motion } from 'framer-motion'
 import { useLazyQuery } from '@apollo/client'
 import { useRouter } from 'next/router'
@@ -8,41 +14,74 @@ import compareAsc from 'date-fns/compareAsc'
 import formatISO from 'date-fns/formatISO'
 
 import {
+  Case,
   CaseDecision,
   CaseState,
+  CaseType,
   InstitutionType,
+  isInvestigationCase,
   isRestrictionCase,
   NotificationType,
   RequestSignatureResponse,
   SignatureConfirmationResponse,
   UserRole,
+  CaseAppealDecision,
+  isAcceptingCaseDecision,
 } from '@island.is/judicial-system/types'
-import type { Case } from '@island.is/judicial-system/types'
 import {
   FormFooter,
   PageLayout,
-  FormContentContainer,
   Modal,
   DateTime,
   BlueBox,
+  InfoCard,
+  PdfButton,
+  CourtRecordAccordionItem,
+  FormContentContainer,
+  PoliceRequestAccordionItem,
+  RulingAccordionItem,
+  CommentsAccordionItem,
+  CaseFilesAccordionItem,
 } from '@island.is/judicial-system-web/src/components'
 import { UserContext } from '@island.is/judicial-system-web/src/components/UserProvider/UserProvider'
-import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
-import { ReactSelectOption } from '@island.is/judicial-system-web/src/types'
-import { Box, Input, Text } from '@island.is/island-ui/core'
-import { FormContext } from '@island.is/judicial-system-web/src/components/FormProvider/FormProvider'
-import { capitalize, formatDate } from '@island.is/judicial-system/formatters'
-import { validate } from '@island.is/judicial-system-web/src/utils/validate'
 import {
+  useCase,
+  useInstitution,
+} from '@island.is/judicial-system-web/src/utils/hooks'
+import { ReactSelectOption } from '@island.is/judicial-system-web/src/types'
+import {
+  Box,
+  Input,
+  Text,
+  Accordion,
+  Button,
+  Select,
+  Tooltip,
+  Stack,
+  Divider,
+  AlertMessage,
+} from '@island.is/island-ui/core'
+import { FormContext } from '@island.is/judicial-system-web/src/components/FormProvider/FormProvider'
+import {
+  capitalize,
+  formatDate,
+  caseTypes,
+} from '@island.is/judicial-system/formatters'
+import { validate } from '@island.is/judicial-system-web/src/utils/validate'
+import MarkdownWrapper from '@island.is/judicial-system-web/src/components/MarkdownWrapper/MarkdownWrapper'
+import PageHeader from '@island.is/judicial-system-web/src/components/PageHeader/PageHeader'
+import * as constants from '@island.is/judicial-system/consts'
+import {
+  core,
   signedVerdictOverview as m,
   titles,
 } from '@island.is/judicial-system-web/messages'
-import MarkdownWrapper from '@island.is/judicial-system-web/src/components/MarkdownWrapper/MarkdownWrapper'
-import PageHeader from '@island.is/judicial-system-web/src/components/PageHeader/PageHeader'
-import * as Constants from '@island.is/judicial-system/consts'
+import { SignedDocument } from '@island.is/judicial-system-web/src/components/SignedDocument/SignedDocument'
+import CaseDates from '@island.is/judicial-system-web/src/components/CaseDates/CaseDates'
+import RestrictionTags from '@island.is/judicial-system-web/src/components/RestrictionTags/RestrictionTags'
 
+import AppealSection from './Components/AppealSection/AppealSection'
 import { CourtRecordSignatureConfirmationQuery } from './courtRecordSignatureConfirmationGql'
-import SignedVerdictOverviewForm from './SignedVerdictOverviewForm'
 
 interface ModalControls {
   open: boolean
@@ -53,6 +92,18 @@ interface ModalControls {
 interface DateTime {
   value?: Date
   isValid: boolean
+}
+
+function showCustodyNotice(
+  type: CaseType,
+  state: CaseState,
+  decision?: CaseDecision,
+) {
+  return (
+    (type === CaseType.CUSTODY || type === CaseType.ADMISSION_TO_FACILITY) &&
+    state === CaseState.ACCEPTED &&
+    isAcceptingCaseDecision(decision)
+  )
 }
 
 export const SignedVerdictOverview: React.FC = () => {
@@ -118,6 +169,61 @@ export const SignedVerdictOverview: React.FC = () => {
     sendNotification,
     isSendingNotification,
   } = useCase()
+
+  // skip loading institutions if the user does not have an id
+  const { prosecutorsOffices } = useInstitution(!user?.id)
+
+  /**
+   * If the case is not rejected it must be accepted because
+   * this screen is only rendered if the case is either accepted
+   * or rejected. Here we are first handling the case where a case
+   * is rejected, then the case where a case is accepted and the
+   * custody end date is in the past and then we assume that
+   * the case is accepted and the custody end date has not come yet.
+   * For accepted cases, we first handle the case where the judge
+   * decided only accept an alternative travel ban and finally we
+   * assume that the actual custody was accepted.
+   */
+  const titleForCase = (theCase: Case) => {
+    const isTravelBan =
+      theCase.decision === CaseDecision.ACCEPTING_ALTERNATIVE_TRAVEL_BAN ||
+      theCase.type === CaseType.TRAVEL_BAN
+
+    if (theCase.state === CaseState.REJECTED) {
+      if (isInvestigationCase(theCase.type)) {
+        return 'Kröfu um rannsóknarheimild hafnað'
+      } else {
+        return 'Kröfu hafnað'
+      }
+    }
+
+    if (theCase.state === CaseState.DISMISSED) {
+      return formatMessage(m.dismissedTitle)
+    }
+
+    if (theCase.isValidToDateInThePast) {
+      return formatMessage(m.validToDateInThePast, {
+        caseType: isTravelBan ? CaseType.TRAVEL_BAN : theCase.type,
+      })
+    }
+
+    return isInvestigationCase(theCase.type)
+      ? formatMessage(m.investigationAccepted)
+      : formatMessage(m.restrictionActive, {
+          caseType: isTravelBan ? CaseType.TRAVEL_BAN : theCase.type,
+        })
+  }
+
+  const canModifyCaseDates = useCallback(() => {
+    return (
+      user &&
+      [UserRole.JUDGE, UserRole.REGISTRAR, UserRole.PROSECUTOR].includes(
+        user.role,
+      ) &&
+      (workingCase.type === CaseType.CUSTODY ||
+        workingCase.type === CaseType.ADMISSION_TO_FACILITY)
+    )
+  }, [workingCase.type, user])
 
   const [getCourtRecordSignatureConfirmation] = useLazyQuery(
     CourtRecordSignatureConfirmationQuery,
@@ -185,19 +291,19 @@ export const SignedVerdictOverview: React.FC = () => {
     if (workingCase) {
       if (workingCase.childCase) {
         if (isRestrictionCase(workingCase.type)) {
-          router.push(`${Constants.STEP_ONE_ROUTE}/${workingCase.childCase.id}`)
+          router.push(`${constants.STEP_ONE_ROUTE}/${workingCase.childCase.id}`)
         } else {
           router.push(
-            `${Constants.IC_DEFENDANT_ROUTE}/${workingCase.childCase.id}`,
+            `${constants.IC_DEFENDANT_ROUTE}/${workingCase.childCase.id}`,
           )
         }
       } else {
         await extendCase(workingCase.id).then((extendedCase) => {
           if (extendedCase) {
             if (isRestrictionCase(extendedCase.type)) {
-              router.push(`${Constants.STEP_ONE_ROUTE}/${extendedCase.id}`)
+              router.push(`${constants.STEP_ONE_ROUTE}/${extendedCase.id}`)
             } else {
-              router.push(`${Constants.IC_DEFENDANT_ROUTE}/${extendedCase.id}`)
+              router.push(`${constants.IC_DEFENDANT_ROUTE}/${extendedCase.id}`)
             }
           }
         })
@@ -272,7 +378,7 @@ export const SignedVerdictOverview: React.FC = () => {
             'dagsins',
           )} kl. ${formatDate(
             modifiedValidToDate?.value,
-            Constants.TIME_FORMAT,
+            constants.TIME_FORMAT,
           )}`,
         },
       )
@@ -287,7 +393,7 @@ export const SignedVerdictOverview: React.FC = () => {
               'dagsins',
             )} kl. ${formatDate(
               modifiedValidToDate?.value,
-              Constants.TIME_FORMAT,
+              constants.TIME_FORMAT,
             )}`,
           },
         )
@@ -302,7 +408,7 @@ export const SignedVerdictOverview: React.FC = () => {
               'PPPP',
             )?.replace('dagur,', 'dagsins')} kl. ${formatDate(
               modifiedIsolationToDate?.value,
-              Constants.TIME_FORMAT,
+              constants.TIME_FORMAT,
             )}`,
           },
         )
@@ -526,7 +632,7 @@ export const SignedVerdictOverview: React.FC = () => {
       workingCase.caseModifiedExplanation ? '<br/><br/>' : ''
     }${capitalize(formatDate(now, 'PPPP', true) || '')} kl. ${formatDate(
       now,
-      Constants.TIME_FORMAT,
+      constants.TIME_FORMAT,
     )} - ${user?.name} ${user?.title}, ${
       user?.institution?.name
     }<br/>Ástæða: ${reason}`
@@ -570,25 +676,367 @@ export const SignedVerdictOverview: React.FC = () => {
       notFound={caseNotFound}
     >
       <PageHeader title={formatMessage(titles.shared.signedVerdictOverview)} />
-      <SignedVerdictOverviewForm
-        workingCase={workingCase}
-        setWorkingCase={setWorkingCase}
-        setAccusedAppealDate={setAccusedAppealDate}
-        setProsecutorAppealDate={setProsecutorAppealDate}
-        withdrawAccusedAppealDate={withdrawAccusedAppealDate}
-        withdrawProsecutorAppealDate={withdrawProsecutorAppealDate}
-        shareCaseWithAnotherInstitution={shareCaseWithAnotherInstitution}
-        selectedSharingInstitutionId={selectedSharingInstitutionId}
-        setSelectedSharingInstitutionId={setSelectedSharingInstitutionId}
-        isRequestingCourtRecordSignature={isRequestingCourtRecordSignature}
-        handleRequestCourtRecordSignature={handleRequestCourtRecordSignature}
-        handleOpenDateModificationModal={() =>
-          setIsModifyingDates(!isModifyingDates)
-        }
-      />
+      <FormContentContainer>
+        <Box marginBottom={5}>
+          <Box marginBottom={3}>
+            <Button
+              variant="text"
+              preTextIcon="arrowBack"
+              onClick={() => router.push(constants.CASE_LIST_ROUTE)}
+            >
+              Til baka
+            </Button>
+          </Box>
+          <Box display="flex" justifyContent="spaceBetween" marginBottom={3}>
+            <Box>
+              <Box marginBottom={1}>
+                <Text as="h1" variant="h1">
+                  {titleForCase(workingCase)}
+                </Text>
+              </Box>
+              <Box>
+                <Text variant="h5">
+                  {formatMessage(m.rulingDateLabel, {
+                    courtEndTime: `${formatDate(
+                      workingCase.courtEndTime,
+                      'PPP',
+                    )} kl. ${formatDate(
+                      workingCase.courtEndTime,
+                      constants.TIME_FORMAT,
+                    )}`,
+                  })}
+                </Text>
+              </Box>
+            </Box>
+            <Box display="flex" flexDirection="column">
+              <RestrictionTags workingCase={workingCase} />
+            </Box>
+          </Box>
+          {isRestrictionCase(workingCase.type) &&
+            workingCase.state === CaseState.ACCEPTED && (
+              <CaseDates
+                workingCase={workingCase}
+                button={
+                  canModifyCaseDates()
+                    ? {
+                        label: formatMessage(core.update),
+                        onClick: () => setIsModifyingDates((value) => !value),
+                        icon: 'pencil',
+                      }
+                    : undefined
+                }
+              />
+            )}
+        </Box>
+        {workingCase.caseModifiedExplanation && (
+          <Box marginBottom={5}>
+            <AlertMessage
+              type="info"
+              title={formatMessage(m.sections.modifyDatesInfo.titleV2, {
+                caseType: workingCase.type,
+              })}
+              message={
+                <MarkdownWrapper
+                  markdown={workingCase.caseModifiedExplanation}
+                  textProps={{ variant: 'small' }}
+                />
+              }
+            />
+          </Box>
+        )}
+        {workingCase.rulingModifiedHistory && (
+          <Box marginBottom={5}>
+            <AlertMessage
+              type="info"
+              title={formatMessage(m.sections.modifyRulingInfo.title)}
+              message={
+                <MarkdownWrapper
+                  markdown={workingCase.rulingModifiedHistory}
+                  textProps={{ variant: 'small' }}
+                />
+              }
+            />
+          </Box>
+        )}
+        <Box marginBottom={6}>
+          <InfoCard
+            data={[
+              {
+                title: formatMessage(core.policeCaseNumber),
+                value: workingCase.policeCaseNumber,
+              },
+              {
+                title: formatMessage(core.courtCaseNumber),
+                value: workingCase.courtCaseNumber,
+              },
+              {
+                title: formatMessage(core.prosecutor),
+                value: `${workingCase.creatingProsecutor?.institution?.name}`,
+              },
+              {
+                title: formatMessage(core.court),
+                value: workingCase.court?.name,
+              },
+              {
+                title: formatMessage(core.prosecutorPerson),
+                value: workingCase.prosecutor?.name,
+              },
+              {
+                title: formatMessage(core.judge),
+                value: workingCase.judge?.name,
+              },
+              // Conditionally add this field based on case type
+              ...(isInvestigationCase(workingCase.type)
+                ? [
+                    {
+                      title: formatMessage(core.caseType),
+                      value: capitalize(caseTypes[workingCase.type]),
+                    },
+                  ]
+                : []),
+              ...(workingCase.registrar
+                ? [
+                    {
+                      title: formatMessage(core.registrar),
+                      value: workingCase.registrar?.name,
+                    },
+                  ]
+                : []),
+            ]}
+            defendants={workingCase.defendants ?? []}
+            defender={{
+              name: workingCase.defenderName ?? '',
+              defenderNationalId: workingCase.defenderNationalId,
+              email: workingCase.defenderEmail,
+              phoneNumber: workingCase.defenderPhoneNumber,
+            }}
+            sessionArrangement={workingCase.sessionArrangements}
+          />
+        </Box>
+        {(workingCase.accusedAppealDecision === CaseAppealDecision.POSTPONE ||
+          workingCase.accusedAppealDecision === CaseAppealDecision.APPEAL ||
+          workingCase.prosecutorAppealDecision ===
+            CaseAppealDecision.POSTPONE ||
+          workingCase.prosecutorAppealDecision === CaseAppealDecision.APPEAL) &&
+          (user?.role === UserRole.JUDGE ||
+            user?.role === UserRole.REGISTRAR) &&
+          user?.institution?.type !== InstitutionType.HIGH_COURT && (
+            <Box marginBottom={7}>
+              <AppealSection
+                workingCase={workingCase}
+                setAccusedAppealDate={setAccusedAppealDate}
+                setProsecutorAppealDate={setProsecutorAppealDate}
+                withdrawAccusedAppealDate={withdrawAccusedAppealDate}
+                withdrawProsecutorAppealDate={withdrawProsecutorAppealDate}
+              />
+            </Box>
+          )}
+        {user?.role !== UserRole.STAFF && (
+          <>
+            <Box marginBottom={5} data-testid="accordionItems">
+              <Accordion>
+                <PoliceRequestAccordionItem workingCase={workingCase} />
+                <CourtRecordAccordionItem workingCase={workingCase} />
+                <RulingAccordionItem workingCase={workingCase} />
+                {user && (
+                  <CaseFilesAccordionItem
+                    workingCase={workingCase}
+                    setWorkingCase={setWorkingCase}
+                    user={user}
+                  />
+                )}
+                {(workingCase.comments ||
+                  workingCase.caseFilesComments ||
+                  workingCase.caseResentExplanation) && (
+                  <CommentsAccordionItem workingCase={workingCase} />
+                )}
+              </Accordion>
+            </Box>
+            <Box marginBottom={6}>
+              <BlueBox>
+                <Box marginBottom={2} textAlign="center">
+                  <Text as="h3" variant="h3">
+                    {formatMessage(m.conclusionTitle)}
+                  </Text>
+                </Box>
+                <Box marginBottom={3}>
+                  <Box marginTop={1}>
+                    <Text variant="intro">{workingCase.conclusion}</Text>
+                  </Box>
+                </Box>
+                <Box marginBottom={1} textAlign="center">
+                  <Text variant="h4">{workingCase?.judge?.name}</Text>
+                </Box>
+              </BlueBox>
+            </Box>
+          </>
+        )}
+        <Box marginBottom={10}>
+          <Text as="h3" variant="h3" marginBottom={3}>
+            {formatMessage(m.caseDocuments)}
+          </Text>
+          <Box marginBottom={2}>
+            <Stack space={2} dividers>
+              {user?.role !== UserRole.STAFF && (
+                <PdfButton
+                  renderAs="row"
+                  caseId={workingCase.id}
+                  title={formatMessage(core.pdfButtonRequest)}
+                  pdfType={'request'}
+                />
+              )}
+              {showCustodyNotice(
+                workingCase.type,
+                workingCase.state,
+                workingCase.decision,
+              ) && (
+                <PdfButton
+                  renderAs="row"
+                  caseId={workingCase.id}
+                  title={formatMessage(core.pdfButtonCustodyNotice)}
+                  pdfType="custodyNotice"
+                />
+              )}
+              <PdfButton
+                renderAs="row"
+                caseId={workingCase.id}
+                title={formatMessage(core.pdfButtonRulingShortVersion)}
+                pdfType={'courtRecord'}
+              >
+                {workingCase.courtRecordSignatory ? (
+                  <SignedDocument
+                    signatory={workingCase.courtRecordSignatory.name}
+                    signingDate={workingCase.courtRecordSignatureDate}
+                  />
+                ) : user?.role === UserRole.JUDGE ||
+                  user?.role === UserRole.REGISTRAR ? (
+                  <Button
+                    variant="ghost"
+                    loading={isRequestingCourtRecordSignature}
+                    onClick={(event) => {
+                      event.stopPropagation()
+                      handleRequestCourtRecordSignature()
+                    }}
+                  >
+                    {formatMessage(m.signButton)}
+                  </Button>
+                ) : (
+                  <Text>{formatMessage(m.unsignedDocument)}</Text>
+                )}
+              </PdfButton>
+              {user?.role !== UserRole.STAFF && (
+                <PdfButton
+                  renderAs="row"
+                  caseId={workingCase.id}
+                  title={formatMessage(core.pdfButtonRuling)}
+                  pdfType={'ruling'}
+                >
+                  <SignedDocument
+                    signatory={workingCase.judge?.name}
+                    signingDate={workingCase.rulingDate}
+                  />
+                  {user?.role === UserRole.JUDGE && (
+                    <Button
+                      variant="ghost"
+                      data-testid="modifyRulingButton"
+                      onClick={(event) => {
+                        event.stopPropagation()
+                        router.push(
+                          isRestrictionCase(workingCase.type)
+                            ? `${constants.MODIFY_RULING_ROUTE}/${workingCase.id}`
+                            : `${constants.IC_MODIFY_RULING_ROUTE}/${workingCase.id}`,
+                        )
+                      }}
+                    >
+                      {capitalize(formatMessage(core.modify))}
+                    </Button>
+                  )}
+                </PdfButton>
+              )}
+            </Stack>
+          </Box>
+          <Divider />
+        </Box>
+        {user?.role === UserRole.PROSECUTOR &&
+          user.institution?.id ===
+            workingCase.creatingProsecutor?.institution?.id &&
+          isRestrictionCase(workingCase.type) && (
+            <Box marginBottom={9}>
+              <Box marginBottom={3}>
+                <Text variant="h3">
+                  {formatMessage(m.sections.shareCase.title)}{' '}
+                  <Tooltip text={formatMessage(m.sections.shareCase.info)} />
+                </Text>
+              </Box>
+              <BlueBox>
+                <Box display="flex">
+                  <Box flexGrow={1} marginRight={2}>
+                    <Select
+                      name="sharedWithProsecutorsOfficeId"
+                      label={formatMessage(m.sections.shareCase.label)}
+                      placeholder={formatMessage(
+                        m.sections.shareCase.placeholder,
+                      )}
+                      size="sm"
+                      icon={
+                        workingCase.sharedWithProsecutorsOffice
+                          ? 'checkmark'
+                          : undefined
+                      }
+                      options={prosecutorsOffices
+                        .map((prosecutorsOffice) => ({
+                          label: prosecutorsOffice.name,
+                          value: prosecutorsOffice.id,
+                        }))
+                        .filter((t) => t.value !== user?.institution?.id)}
+                      value={
+                        workingCase.sharedWithProsecutorsOffice
+                          ? {
+                              label:
+                                workingCase.sharedWithProsecutorsOffice.name,
+                              value: workingCase.sharedWithProsecutorsOffice.id,
+                            }
+                          : selectedSharingInstitutionId
+                          ? {
+                              label: (selectedSharingInstitutionId as ReactSelectOption)
+                                .label,
+                              value: (selectedSharingInstitutionId as ReactSelectOption)
+                                .value as string,
+                            }
+                          : null
+                      }
+                      onChange={(so: ValueType<ReactSelectOption>) =>
+                        setSelectedSharingInstitutionId(so)
+                      }
+                      disabled={Boolean(
+                        workingCase.sharedWithProsecutorsOffice,
+                      )}
+                    />
+                  </Box>
+                  <Button
+                    size="small"
+                    disabled={
+                      !selectedSharingInstitutionId &&
+                      !workingCase.sharedWithProsecutorsOffice
+                    }
+                    onClick={() =>
+                      shareCaseWithAnotherInstitution(
+                        selectedSharingInstitutionId,
+                      )
+                    }
+                  >
+                    {workingCase.sharedWithProsecutorsOffice
+                      ? formatMessage(m.sections.shareCase.close)
+                      : formatMessage(m.sections.shareCase.open)}
+                  </Button>
+                </Box>
+              </BlueBox>
+            </Box>
+          )}
+      </FormContentContainer>
       <FormContentContainer isFooter>
         <FormFooter
-          previousUrl={Constants.CASE_LIST_ROUTE}
+          previousUrl={constants.CASE_LIST_ROUTE}
           hideNextButton={
             user?.role !== UserRole.PROSECUTOR ||
             workingCase.decision ===
